ECU-10143 Refactor initial implementation
