+++ README.md
@@ -1,1 +1,25 @@
-# upgrade-startup-deployer
+# UPGRADE STARTUP DEPLOYER
+
+The Module Deployer handles the deployment of all necessary submodules required by the module you intend to deploy. This ensures that, in OSGi, the bundle generated by your module is deployed only after all sub-bundles upon which it depends have been successfully deployed.
+
+Note: So far, this tool only supports Gradle projects
+
+## How to build and run the project
+
+### Requirements
+
+**JDK 21**
+
+1 - To build the project from source, run:
+
+```
+./gradlew clean build
+```
+
+To deployer an upgrade project using your local build run:
+
+```
+java -jar build/libs/upgrade-startup-*.jar -md -f /path/to/client/workspace
+```
+
+For more details about this project, refer to the documentation available in Confluence: [Module deployer](https://liferay.atlassian.net/wiki/spaces/ECU/pages/3676668235/Module+deployer?force_transition=3f17c7e0-db73-48a7-89a6-003d039d48a6)
