+++ build.gradle
@@ -1,0 +1,44 @@
+plugins {
+    id 'java'
+}
+
+group = 'com.liferay.upgrades'
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation group: "org.reflections", name: "reflections", version: "0.10.2"
+    implementation 'org.apache.commons:commons-lang3:3.13.0'
+    implementation group: 'org.json', name: 'json', version: '20240303'
+    implementation group: 'org.apache.logging.log4j', name: 'log4j-api', version: '2.19.0'
+    implementation group: 'org.apache.logging.log4j', name: 'log4j-core', version: '2.19.0'
+
+    testImplementation platform('org.junit:junit-bom:5.9.1')
+    testImplementation 'org.junit.jupiter:junit-jupiter'
+
+    implementation group: "com.beust", name: "jcommander", version: "1.82"
+}
+
+java {
+    toolchain {
+        languageVersion = JavaLanguageVersion.of(21)
+    }
+}
+
+jar {
+    duplicatesStrategy = DuplicatesStrategy.INCLUDE
+    manifest {
+        attributes(
+                'Main-Class': 'com.liferay.upgrades.startup.main.Main'
+        )
+    }
+    from {
+        configurations.runtimeClasspath.collect { it.isDirectory() ? it : zipTree(it) }
+    }
+}
+
+test {
+    useJUnitPlatform()
+}
