+++ src/main/java/com/liferay/upgrades/startup/main/Main.java
@@ -1,14 +1,14 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package com.liferay.upgrades.startup;
+package com.liferay.upgrades.startup.main;
 
-public class App {
+public class Main {
     public String getGreeting() {
         return "Hello World!";
     }
 
     public static void main(String[] args) {
-        System.out.println(new App().getGreeting());
+        System.out.println(new Main().getGreeting());
     }
 }
