+++ src/test/java/com/liferay/upgrades/startup/MainTest.java
@@ -3,12 +3,13 @@
  */
 package com.liferay.upgrades.startup;
 
+import com.liferay.upgrades.startup.main.Main;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
 
-class AppTest {
+class MainTest {
     @Test void appHasAGreeting() {
-        App classUnderTest = new App();
+        Main classUnderTest = new Main();
         assertNotNull(classUnderTest.getGreeting(), "app should have a greeting");
     }
 }
